==> notebooks/Model Pre-processing/1. Cleaning.ipynb <==
import pandas as pd
import numpy as np
import glob
import sys
sys.path.append('../..')
from helpers import lookups

pd.set_option('display.max_columns', None)

dfs = [pd.read_csv(file, parse_dates=['fecha']) for file in glob.glob('../../data/historical/historico/*.csv')]

df = pd.concat(dfs, ignore_index=True)
df = df.sort_values(by='fecha').reset_index(drop=True)

nan_percentage = df.isna().mean() * 100
nan_percentage = nan_percentage[nan_percentage > 0].sort_values(ascending=False)
nan_percentage.apply(lambda x: f"{x:.2f}%")

import sys
sys.path.append('..')
from helpers import lookups


def clean_df(df: pd.DataFrame) -> pd.DataFrame:
    df_clean = df.copy()
    
    df_clean = df_clean.drop(columns=lookups.time_cols)
    
    df_clean = df_clean.merge(lookups.locations_df[['idema', 'latitud', 'longitud', 'altitud']], on='idema', how='left')
    df_clean['fecha_day'] = df_clean['fecha'].dt.dayofyear
    
    df_clean['fecha_sin'] = df_clean.apply(
        lambda row: (np.sin(2 * np.pi * row['fecha_day'] / 366) + 1) / 2 if row['year'] % 4 == 0 
        else (np.sin(2 * np.pi * row['fecha_day'] / 365) + 1) / 2, axis=1)
    
    df_clean['fecha_cos'] = df_clean.apply(
        lambda row: (np.cos(2 * np.pi * row['fecha_day'] / 366) + 1) / 2 if row['year'] % 4 == 0 
        else (np.cos(2 * np.pi * row['fecha_day'] / 365) + 1) / 2, axis=1)

    numeric_cols = df_clean.drop(columns=['fecha', 'idema']).columns
    df_clean[numeric_cols] = df_clean[numeric_cols].apply(pd.to_numeric, errors='coerce')

    

df_clean = df.copy()

df_clean = df_clean.drop(columns=lookups.time_cols)

df_clean = df_clean.merge(lookups.locations_df[['idema', 'latitud', 'longitud', 'altitud']], on='idema', how='left')

df_clean['year'] = df_clean['fecha'].dt.year
df_clean['year'] = (df_clean['year'] - 1950) / 100

df_clean['fecha_day'] = df_clean['fecha'].dt.dayofyear

df_clean['fecha_sin'] = df_clean.apply(
    lambda row: (np.sin(2 * np.pi * row['fecha_day'] / 366) + 1) / 2 if row['year'] % 4 == 0 
    else (np.sin(2 * np.pi * row['fecha_day'] / 365) + 1) / 2, axis=1)

df_clean['fecha_cos'] = df_clean.apply(
    lambda row: (np.cos(2 * np.pi * row['fecha_day'] / 366) + 1) / 2 if row['year'] % 4 == 0 
    else (np.cos(2 * np.pi * row['fecha_day'] / 365) + 1) / 2, axis=1)

numeric_cols = df_clean.drop(columns=['fecha', 'idema']).columns
df_clean[numeric_cols] = df_clean[numeric_cols].apply(pd.to_numeric, errors='coerce')

df_clean['dir_sin'] = df_clean.apply(lambda row: np.sin(2 * np.pi * row['dir'] / 99), axis=1)
df_clean['dir_cos'] = df_clean.apply(lambda row: np.cos(2 * np.pi * row['dir'] / 99), axis=1)

df_clean = df_clean.drop(columns=['dir', 'fecha_day'])

df_clean = df_clean.round(4)

df_clean.head()

df_clean.to_csv('../../data/ml/clean.csv', index=False)
