==> notebooks/Model Pre-processing/3. Closest Stations Lookup.ipynb <==
import pandas as pd
import numpy as np
import json
from collections import OrderedDict
from sklearn.neighbors import NearestNeighbors
from sklearn.impute import KNNImputer
from sklearn.preprocessing import StandardScaler

estaciones = pd.read_csv('../data/locations/estaciones.csv')

estaciones = estaciones.drop(columns=['nombre', 'provincia_id'])

estaciones.head()

knn_features = ['latitud', 'longitud', 'altitud']

scaler = StandardScaler()

estaciones[knn_features] = scaler.fit_transform(estaciones[knn_features].values)

estaciones.head()

nearest_neighbors = NearestNeighbors(n_neighbors=len(estaciones), metric='euclidean')

nearest_neighbors.fit(estaciones[knn_features].values)

distances, indices = nearest_neighbors.kneighbors(estaciones[knn_features].values)

lookup = {}

for i, row in estaciones.iterrows():
    nearest_indices = indices[i]
    closest_idemas = estaciones.iloc[nearest_indices]['idema'].tolist()
    lookup[row['idema']] = closest_idemas

for i in range(3):
    display(estaciones[estaciones['idema'] == lookup['0016A'][i]])

lookup_df = pd.DataFrame.from_dict(lookup)
lookup_df = lookup_df.drop(0).reset_index(drop=True) # Ignore itself

lookup_df.head()

lookup_df.to_csv('../data/ml/estacion_distance_lookup.csv', index=False)
