==> notebooks/Model Pre-processing/4. Impute by Nearest Station.ipynb <==
import pandas as pd
import time

df = pd.read_csv('../../data/ml/filtered.csv', parse_dates=['fecha'], low_memory=False)

df.head()

nan_percentage = df.isna().mean() * 100
nan_percentage = nan_percentage[nan_percentage > 0].sort_values(ascending=False)
nan_percentage.apply(lambda x: f"{x:.2f}%")

lookup = pd.read_csv('../../data/locations/estacion_distance_lookup.csv')

df_imputed = df.copy()

def impute_knn(col, k=3):

    # Iterate through each day
    for fecha in df_imputed['fecha'].unique():
        current_row = df_imputed[df_imputed['fecha'] == fecha]

        # Filter NaN and non NaN rows
        nan_rows = current_row[current_row[col].isna()]
        non_nan_rows = current_row[current_row[col].notna()]

        for index, row in nan_rows.iterrows():
            
            # Use the nearest k stations to take the average
            nearest_idemas = lookup[row['idema']]
            nearest_values = non_nan_rows[non_nan_rows['idema'].isin(nearest_idemas)][col][:k]

            if not nearest_values.empty:
                imputed_value = nearest_values.mean()
                df_imputed.loc[index, col] = imputed_value

Since the datasets will be split by year, this does not introduce any data leakage as it goes day by day

start = time.time()
for col in ['tmed', 'tmax', 'tmin', 'prec', 'hr_media', 'hr_max']:
    col_start = time.time()
    impute_knn(col)
    elapsed = time.time() - col_start
    print(f"Imputed {col} in {elapsed}s")
print(f"Imputed all columns in {time.time() - start}s")

df_imputed.head()

df_imputed.info(show_counts=True)

df_imputed.to_csv('../../data/ml/full.csv', index=False)
