==> notebooks/Model Pre-processing/5. Train-Validation-Test Split.ipynb <==
import pandas as pd

df = pd.read_csv('../../data/ml/full.csv', parse_dates=['fecha'])

df.info(show_counts=True)

train_df = df[df['fecha'].dt.year <= 2020]
validation_df = df[(df['fecha'].dt.year >= 2021) & (df['fecha'].dt.year <= 2022)]
test_df = df[df['fecha'].dt.year >= 2023]

len(train_df), len(validation_df), len(test_df)

round(len(train_df) / len(df) * 100, 2), round(len(validation_df) / len(df) * 100, 2), round(len(test_df) / len(df) * 100, 2)

train_df.head(1)

validation_df.head(1)

test_df.head(1)

train_df.to_csv('../../data/ml/train.csv', index=False)
validation_df.to_csv('../../data/ml/validation.csv', index=False)
test_df.to_csv('../../data/ml/test.csv', index=False)
