==> notebooks/Model Pre-processing/6. Scaling.ipynb <==
import pandas as pd
import joblib
from sklearn.preprocessing import MinMaxScaler

full_df = pd.read_csv('../../data/ml/full.csv', parse_dates=['fecha'])
train_df = pd.read_csv('../../data/ml/train.csv', parse_dates=['fecha'])
validation_df = pd.read_csv('../../data/ml/validation.csv', parse_dates=['fecha'])
test_df = pd.read_csv('../../data/ml/test.csv', parse_dates=['fecha'])

validation_df['tmed'].head()

scaler_X_full = MinMaxScaler(feature_range=(0, 1))
scaler_X_train = MinMaxScaler(feature_range=(0, 1))
scaler_y_full = MinMaxScaler(feature_range=(0, 1))
scaler_y_train = MinMaxScaler(feature_range=(0, 1))

def scale_df(df: pd.DataFrame, scaler_X, scaler_y, fit: bool = False) -> pd.DataFrame:

    df_unscaled = df.copy()
    
    target = 'tmed'
    ignore_cols = ['fecha', 'idema', 'fecha_sin', 'fecha_cos']
    
    X_scale_cols = list(df_unscaled.drop(columns=ignore_cols + [target]).columns)
    print(X_scale_cols)
    y_scale_cols = [target]

    if fit:
        X_scaled = scaler_X.fit_transform(df_unscaled[X_scale_cols].values)
        y_scaled = scaler_y.fit_transform(df_unscaled[y_scale_cols].values)
    else:
        X_scaled = scaler_X.transform(df_unscaled[X_scale_cols].values)
        y_scaled = scaler_y.transform(df_unscaled[y_scale_cols].values)
    
    df_scaled = pd.DataFrame(X_scaled, columns=df_unscaled[X_scale_cols].columns)
    df_scaled[target] = y_scaled

    df_scaled[ignore_cols] = df_unscaled[ignore_cols].values
    df_scaled = df_scaled[df_unscaled.columns]

    numeric_cols = ['fecha_sin', 'fecha_cos']
    df_scaled[numeric_cols] = df_scaled[numeric_cols].apply(pd.to_numeric, errors='coerce')

    return df_scaled.round(4)

full_df_scaled = scale_df(full_df, scaler_X_full, scaler_y_full, fit=True)
train_df_scaled = scale_df(train_df, scaler_X_train, scaler_y_train, fit=True)

joblib.dump(scaler_X_full, '../../ml/scalers/scaler_X_full.joblib')
joblib.dump(scaler_y_full, '../../ml/scalers/scaler_y_full.joblib')
joblib.dump(scaler_X_train, '../../ml/scalers/scaler_X_train.joblib')
joblib.dump(scaler_y_train, '../../ml/scalers/scaler_y_train.joblib')

validation_df_scaled = scale_df(validation_df, scaler_X_train, scaler_y_train, fit=False)
test_df_scaled = scale_df(test_df, scaler_X_train, scaler_y_train, fit=False)

full_df_scaled.head(3)

train_df_scaled.head(3)

validation_df_scaled.head(3)

test_df_scaled.head(3)

numeric_cols = full_df_scaled.select_dtypes(include='number')
numeric_cols.max() - numeric_cols.min()

full_df_scaled.to_csv('../../data/ml/full_scaled.csv', index=False)
train_df_scaled.to_csv('../../data/ml/train_scaled.csv', index=False)
validation_df_scaled.to_csv('../../data/ml/validation_scaled.csv', index=False)
test_df_scaled.to_csv('../../data/ml/test_scaled.csv', index=False)
